security fix

+++ routes/report.js
@@ -3,20 +3,22 @@ import Report from "../models/Report.js";
 import authMiddleware from "../middlewares/authMiddleware.js";
 import sanitizeHtml from "sanitize-html";
 import User from "../models/User.js";
+import ExpressMongoSanitize from "express-mongo-sanitize";
 
 const router = express.Router();
 
 router.post("/", authMiddleware, async (req, res) => {
     try {
-        const { target, targetType, targetPost ,reason, description } = req.body;
+        const { target, targetType, targetPost ,reason, description } = ExpressMongoSanitize.sanitize(req.body); 
         if (!target || !targetType || !reason) {
             return res.status(400).json({ message: "Missing required fields" });
         }
 
         const existingReport = await Report.findOne({
-            reporter: req.user.userID,
-            target: target
+            reporter: String(req.user.userID),
+            target: String(target)
         })
+        console.log(typeof(req.user.userID))
         if (existingReport) return res.status(409).json({ message: "You have already reported this content" })
 
         if (description.length > 300) return res.status(413).json({ message: "Description is too long" });
@@ -64,15 +66,18 @@ router.get("/", authMiddleware, async (req,res)=>{
 router.delete('/:id', authMiddleware, async (req, res) => {
     try {
         const currentUser = await User.findById(req.user.userID);
+        if (!currentUser) {
+            return res.status(404).json({ message: "User not found" });
+        }
         if(currentUser.role !== "admin") {
-            return res.status(401).json({message : "Unauth"})
+            return res.status(403).json({message : "Unauth"})
         }
         else{
             await Report.findByIdAndDelete(req.params.id);
             res.status(200).json("Report resolved/deleted");
         }
     } catch (err) {
-        res.status(500).json(err);
+        res.status(500).json({error : err.message});
     }
 });
